# gateway/src/main/java/ru/practicum/shareit/booking/dto/BookItemRequestDto.java
package ru.practicum.shareit.booking.dto;

import java.time.LocalDateTime;

import javax.validation.constraints.Future;
import javax.validation.constraints.FutureOrPresent;
import javax.validation.constraints.NotNull;

import lombok.*;

@Value
@Builder(toBuilder = true)
@AllArgsConstructor
public class BookItemRequestDto {

    long id;

    long itemId;

    @NotNull
    @FutureOrPresent
    LocalDateTime start;

    @NotNull
    @Future
    LocalDateTime end;
}