# gateway/src/main/java/ru/practicum/shareit/item/dto/CommentDtoAdd.java
package ru.practicum.shareit.item.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Value;
import ru.practicum.shareit.user.dto.UserDto;

import javax.validation.constraints.NotBlank;
import java.time.LocalDateTime;

@Value
@Builder(toBuilder = true)
@AllArgsConstructor
public class CommentDtoAdd {

    Long id;

    @NotBlank
    String text;

    ItemDto item;

    UserDto author;

    LocalDateTime created;
}