# gateway/src/main/java/ru/practicum/shareit/user/dto/UserDto.java
package ru.practicum.shareit.user.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Value;

@Value
@Builder(toBuilder = true)
@AllArgsConstructor
public class UserDto {

    Long id;

    String name;

    String email;
}